requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
coverage==6.2
Django==4.0
django-countries==7.2.1
django-extensions==3.1.5
django-stubs==1.9.0
django-stubs-ext==0.3.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
six==1.16.0
sqlparse==0.4.2
stripe3==0.0.0
toml==0.10.2
tomli==1.2.3
types-pytz==2021.3.2
types-PyYAML==6.0.1
typing_extensions==4.0.1
